Add audio playback toggle to settings
