+++ app/src/main/res/menu/drawer_menu.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/nav_settings"
+        android:title="@string/settings_title"
+        android:icon="@android:drawable/ic_menu_preferences" />
+</menu>
